+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2022-12-12
+
 ### Added
 
 - Support for adding team reviewers by setting `team_reviewers`
@@ -37,7 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release!
 
-[Unreleased]: https://github.com/thomaseizinger/create-pull-request/compare/1.2.0...HEAD
+[Unreleased]: https://github.com/thomaseizinger/create-pull-request/compare/1.3.0...HEAD
+[1.3.0]: https://github.com/thomaseizinger/create-pull-request/compare/1.2.0...1.3.0
 [1.2.0]: https://github.com/thomaseizinger/create-pull-request/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/thomaseizinger/create-pull-request/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/thomaseizinger/create-pull-request/compare/92284b92aff90f2100e022ed93d6e485240e8a36...1.0.0